+++ src/core/utils/index.js
@@ -6,6 +6,7 @@ import * as dom from './dom.js';
 import * as html from './html.js';
 import * as json from './json.js';
 import * as layout from './layout.js';
+import * as pattern from './pattern.js';
 import * as persistent from './persistent.js';
 import safeFilterRx from './safe-filter-rx.js';
 
@@ -18,6 +19,7 @@ export default {
     ...html,
     ...json,
     ...layout,
+    pattern,
     persistent,
     safeFilterRx
 };