+++ src/views/text-match.usage.js
@@ -22,7 +22,53 @@ export default {
             demo: {
                 view: 'link',
                 data: { href: '#' },
-                content: 'text-match:{ text: "Test link testtest", match: /test/i }'
+                content: 'text-match:{ text: "Test link", match: /\\w+/i }'
+            }
+        },
+        {
+            title: 'Various configurations',
+            demo: {
+                view: 'table',
+                cols: [
+                    { header: 'text-match', content: 'text-match' },
+                    { header: 'Config', content: 'struct' }
+                ],
+                data: [
+                    null,
+                    { },
+                    { },
+                    {
+                        match: 'test'
+                    },
+                    {
+                        match: 'Test'
+                    },
+                    {
+                        match: 'test',
+                        ignoreCase: true
+                    },
+                    {
+                        match: 'Test',
+                        ignoreCase: true
+                    },
+                    {
+                        match: /test/
+                    },
+                    {
+                        match: /test/g
+                    },
+                    {
+                        match: /test/i
+                    },
+                    {
+                        match: /test/,
+                        ignoreCase: true
+                    },
+                    {
+                        match: /test/g,
+                        ignoreCase: true
+                    }
+                ].map((item, idx) => idx > 1 ? { text: 'test Test', ...item } : item)
             }
         }
     ]