+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `onClick` option for link view
 - Added `ignoreCase` option to `text-match` view
 - Fixed various edge cases for `text-match` view
+- Fixed markup for `tree-leaf` view in non-collapsible state
 - Exposed internal helpers as `utils`: `debounce`, `jsonStringifyInfo`, `jsonStringifyAsJavaScript`, `persistent` and `pattern` modules
 
 ## 1.0.0-beta.46 (23-10-2020)