+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - Fixed various edge cases for `text-match` view
 - Fixed markup for `tree-leaf` view in non-collapsible state
 - Fixed data for `toggle` view when used in `toggle-group`
-- Exposed internal helpers as `utils`: `debounce`, `jsonStringifyInfo`, `jsonStringifyAsJavaScript`, `persistent`, `pattern` and `pointerXY` modules
+- Removed margin reset for header views when first child of block
+- Exposed internal helpers as `utils`: `debounce`, `contentRect`, `jsonStringifyInfo`, `jsonStringifyAsJavaScript`, `persistent`, `pattern` and `pointerXY` modules
 
 ## 1.0.0-beta.46 (23-10-2020)